+++ R/brglmFit.R
@@ -69,14 +69,16 @@
 #' iteration with the same stationary point.
 #'
 #' In the special case of generalized linear models for binomial,
-#' Poisson and multinomial responses, the adjusted score equations
-#' approach returns estimates with improved frequentist properties,
-#' that are also always finite, even in cases where the maximum
-#' likelihood estimates are infinite (e.g. complete and quasi-complete
-#' separation in multinomial regression). See, Kosmidis and Firth
-#' (2021) for a proof for binomial-response GLMs with Jeffreys-prior
-#' penalties to the log-likelihood, which is equivalent to mean bias
-#' reduction for logistic regression. See, also,
+#' Poisson and multinomial responses, the adjusted score equation
+#' approaches for \code{type = "AS_mixed"}, \code{type = "AS_mean"},
+#' and \code{type = "AS_median"} (see below for what methods each
+#' \code{type} corresponds) return estimates with improved frequentist
+#' properties, that are also always finite, even in cases where the
+#' maximum likelihood estimates are infinite (e.g. complete and
+#' quasi-complete separation in multinomial regression). See, Kosmidis
+#' and Firth (2021) for a proof for binomial-response GLMs with
+#' Jeffreys-prior penalties to the log-likelihood, which is equivalent
+#' to mean bias reduction for logistic regression. See, also,
 #' \code{\link[detectseparation]{detect_separation}} and
 #' \code{\link[detectseparation]{check_infinite_estimates}} for
 #' pre-fit and post-fit methods for the detection of infinite
@@ -1031,7 +1033,7 @@ brglmFit <- function(x, y, weights = rep(1, nobs), start = NULL, etastart = NULL
 
         ## Convergence analysis
         if ((failed | iter >= control$maxit) & !(is_correction)) {
-            warning("brglmFit: algorithm did not converge", call. = FALSE)
+            warning("brglmFit: algorithm did not converge. Try changing the optimization algorithm defaults, e.g. the defaults for one or more of `maxit`, `epsilon`, `slowit`, and `response_adjustment`; see `?brglm_control` for default values and available options", call. = FALSE)
             converged <- FALSE
         }
         else {
@@ -1274,9 +1276,12 @@ summary.brglmFit <- function(object, dispersion = NULL,
             dispersion <- object$dispersion
         }
     }
-    summary.glm(object, dispersion = dispersion,
-                correlation = correlation,
-                symbolic.cor = symbolic.cor, ...)
+    out <- summary.glm(object, dispersion = dispersion,
+                       correlation = correlation,
+                       symbolic.cor = symbolic.cor, ...)
+    out$type <- object$type
+    class(out) <- c("summary.brglmFit", class(out))
+    out
 }
 
 #' Method for computing confidence intervals for one or more
@@ -1338,3 +1343,69 @@ DD <- function(expr,name, order = 1) {
 
 
 
+## Almost all code is from stats:::print.summary.glm apart from minor modifications
+#' @rdname summary.brglmFit
+#' @method print summary.brglmFit
+#' @export
+print.summary.brglmFit <- function (x, digits = max(3L, getOption("digits") - 3L),
+                                    symbolic.cor = x$symbolic.cor,
+                                    signif.stars = getOption("show.signif.stars"), ...) {
+    cat("\nCall:\n", paste(deparse(x$call), sep = "\n", collapse = "\n"),
+        "\n\n", sep = "")
+    cat("Deviance Residuals: \n")
+    if (x$df.residual > 5) {
+        x$deviance.resid <- setNames(quantile(x$deviance.resid,
+            na.rm = TRUE), c("Min", "1Q", "Median", "3Q", "Max"))
+    }
+    xx <- zapsmall(x$deviance.resid, digits + 1L)
+    print.default(xx, digits = digits, na.print = "", print.gap = 2L)
+    if (length(x$aliased) == 0L) {
+        cat("\nNo Coefficients\n")
+    }
+    else {
+        df <- if ("df" %in% names(x))
+            x[["df"]]
+        else NULL
+        if (!is.null(df) && (nsingular <- df[3L] - df[1L]))
+            cat("\nCoefficients: (", nsingular, " not defined because of singularities)\n",
+                sep = "")
+        else cat("\nCoefficients:\n")
+        coefs <- x$coefficients
+        if (!is.null(aliased <- x$aliased) && any(aliased)) {
+            cn <- names(aliased)
+            coefs <- matrix(NA, length(aliased), 4L, dimnames = list(cn,
+                colnames(coefs)))
+            coefs[!aliased, ] <- x$coefficients
+        }
+        printCoefmat(coefs, digits = digits, signif.stars = signif.stars,
+            na.print = "NA", ...)
+    }
+    cat("\n(Dispersion parameter for ", x$family$family, " family taken to be ",
+        format(x$dispersion), ")\n\n", apply(cbind(paste(format(c("Null",
+            "Residual"), justify = "right"), "deviance:"), format(unlist(x[c("null.deviance",
+            "deviance")]), digits = max(5L, digits + 1L)), " on",
+            format(unlist(x[c("df.null", "df.residual")])), " degrees of freedom\n"),
+            1L, paste, collapse = " "), sep = "")
+    if (nzchar(mess <- naprint(x$na.action)))
+        cat("  (", mess, ")\n", sep = "")
+    cat("AIC: ", format(x$aic, digits = max(4L, digits + 1L)))
+    cat("\n\nType of estimator:", x$type, get_type_description(x$type))
+    cat("\n", "Number of Fisher Scoring iterations: ", x$iter, "\n", sep = "")
+    correl <- x$correlation
+    if (!is.null(correl)) {
+        p <- NCOL(correl)
+        if (p > 1) {
+            cat("\nCorrelation of Coefficients:\n")
+            if (is.logical(symbolic.cor) && symbolic.cor) {
+                print(symnum(correl, abbr.colnames = NULL))
+            }
+            else {
+                correl <- format(round(correl, 2L), nsmall = 2L,
+                  digits = digits)
+                correl[!lower.tri(correl)] <- ""
+                print(correl[-1, -p, drop = FALSE], quote = FALSE)
+            }
+        }
+    }
+    invisible(x)
+}